+++ week-2/01-async-js/easy/2-counter.js
@@ -1,0 +1,83 @@
+// ## Counter without setInterval
+
+// Without using setInterval, try to code a counter in Javascript. There is a hint at the bottom of the file if you get stuck.
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+// (Hint: setTimeout)
+let timer =0 ;
+function updateTimer(){
+    timer++;
+    console.log(timer);
+    setTimeout(updateTimer,1000);
+}
+updateTimer()
