--- week-2/01-async-js/easy/2-counter.md
@@ -1,76 +1,0 @@
-## Counter without setInterval
-
-Without using setInterval, try to code a counter in Javascript. There is a hint at the bottom of the file if you get stuck.
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-(Hint: setTimeout)
