--- week-2/01-async-js/easy/4-write-to-file.md
@@ -1,3 +1,0 @@
-## Write to a file
-Using the fs library again, try to write to the contents of a file.
-You can use the fs library to as a black box, the goal is to understand async tasks.
