--- week-2/01-async-js/easy/counter.js
@@ -1,6 +1,0 @@
-let i = 0;
-function counter(){
-    console.log(i);
-    i++;
-}
-setInterval(counter,1000)
