--- week-2/01-async-js/easy/counter1.js
@@ -1,7 +1,0 @@
-let timer =0 ;
-function updateTimer(){
-    timer++;
-    console.log(timer);
-    setTimeout(updateTimer,1000);
-}
-updateTimer()
